Example Server Code I Have Been Working On
Currently all it does is display the phaser logo and keep track
of the amount of players connected. The amount of players
connected gets updated in the log.

To run navigate into exampleCode directory and run node server.js
and then navigate to http://compute.cse.tamu.edu:31337 in your
browser.